% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_create_single_bm.R
\name{create_single_bm}
\alias{create_single_bm}
\title{Create a single mlr3 benchmark run including tuning}
\usage{
create_single_bm(
  learner,
  learner_id,
  task,
  measure,
  resampling_inner,
  resampling_outer,
  search_space,
  term_evals,
  term_time = 7 * 24 * 60 * 60,
  method,
  fallback = NULL,
  store_models = TRUE
)
}
\description{
Creates an mlr3 learner which performs hyperparameter tuning in nested resampling
}
